pyproject.toml
[project]
name = "rag-document-qa"
version = "1.0.0"
description = "RAG-based Document Q&A System with Personal Watermark Protection"
authors = [
    {name = "Balenci Cash", email = "balencicash@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langgraph>=0.0.20",
    "langsmith>=0.0.70",
    "openai>=1.6.0",
    "faiss-cpu>=1.7.4",
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.6",
    "markdown>=3.5",
    "loguru>=0.7.2",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.2",
    "numpy>=1.24.0",
    "cryptography>=41.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

# Personal Watermark Configuration
# Copyright (c) 2024 Balenci Cash - All Rights Reserved
# This code is protected and watermarked for authentication purposes
# Unauthorized use or distribution is prohibited